refactor(cli): restart -> chokidar
